feat(computer-use): emit typed TerminationReason enum in StopRecordingResult (REMOTE-2165)
Replace the free-form string termination_reason field with a typed
RecordingTerminationReason enum that mirrors StopRecordingResult.TerminationReason
in warp-proto-apis (PR warpdotdev/warp-proto-apis#348).

Changing from string (wire type 2) to enum (wire type 0) is a breaking
proto change. Without this client update the field would be silently
dropped by the server, losing termination-reason fidelity in the
recording telemetry dashboard.

Changes:
- Bump warp_multi_agent_api rev to db991de (adds TerminationReason enum)
- Add RecordingTerminationReason enum to ai::agent::action_result
- Change RecordingStopped.termination_reason: String -> RecordingTerminationReason
- Update FinalizeReason::termination_reason() to return the enum
- Update convert.rs to cast enum as i32 for the proto wire field
- Add convert_recording_termination_reason() to convert_conversation.rs
- Update tests to use RecordingTerminationReason enum values

LimitReached is mapped to MaxDuration as a best-effort approximation;
the client does not currently distinguish duration vs. size limits at
the RecordingExitKind level.

+++ crates/ai/src/agent/action_result/convert.rs
@@ -1478,7 +1478,7 @@ impl TryFrom<StopRecordingResult> for api::request::input::tool_call_result::Res
                                 completion_status: convert_completion_status(
                                     stopped.completion_status,
                                 ) as i32,
-                                termination_reason: stopped.termination_reason,
+                                termination_reason: stopped.termination_reason as i32,
                             },
                         )),
                     },
